requirements.txt
# closechimp required pip libraries

postmonkey==1.0b
requests==2.5.1
requests-transition==1.0.4.0